requirements.txt
pandas==2.1.2
matplotlib==3.8.1
numpy==1.26.1
python-dotenv==1.0.0